move all files into a upper folder
